+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **Changed**
   * LED restructure set/get pixel functions to simplify array subclassing.
   * LED some minor cleanup of array subtypes.
+  * MQTT use custom client logger for better control of log messages.
 
 * **Fixed**
   * Missing comma in Odroid C1+ and C2 board definitions.